# Optimize concat, crop, pad