--- a/augly/video/augmenters/ffmpeg/crop.py
+++ b/augly/video/augmenters/ffmpeg/crop.py
@@ -27,8 +27,8 @@ class VideoAugmenterByCrop(BaseVidgearFFMPEGAugmenter):
 
         @param output_path: the path in which the resulting video will be stored.
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         video_info = get_video_info(video_path)